fix: catch unhandled errors in get_all_source function
If an app or tool uses this method to fetch sources and there is an invalid source, it will fail when the source is detected. This changes this behavior to instead print the error to console, and optionally fail with the exception, allowing tools to decide if they should failover hard or continue to display the rest of the sources.