fix(source): add comments to list files
Adds a comment attribute to source objects and allows legacy deb files to have comments. This allows files added by third parties with comments in the file to have the comments preserved after Repolib touches a file. Particularly useful with files like those in #32 where the comment describes the behavior of the repo.