fix(debian): improve preinst and prerm behavior
Rewrite the preinst script (and add a prerm script) that will
improve fault tolerance for certain situations that may be
encountered depending on packages and user-setup on systems. This
ensures that we only create a symlink for add-apt-repository if
it isn't already present, that we don't attempt to create a
duplicate (and fail) if we're already managing it, and we don't
remove them unless we are the ones who set it up.

This also splits handling of both add-apt-repository and
apt-add-repository into separate sections so that we can handle
them individually.