feat(cli/remove): offer suggestions for removing sources
We build a suggested source by passing the source the user entered through the
ident-generation. If the suggested source is in that list, offer this as a
suggested removal instead.

Note: This bypasses the -y flag because we don't want to automatically
remove the source if it needed correction. Subsquent prompts will still
follow the -y prompts.