fix(cli/add): catch exceptions from adding sources
Common errors include being offline for popdev and ppa sources. We also catch
any generic exception and print the error to console. Any time an exception is
caught which can't be handled, we quit with 1. This ensures that scripts
can handle the error correctly, and also avoid having an overwhelming
traceback from being printed to console.