feat(source): add method to return signing key info
This returns the first item in the gnupg keyring returned by the
list_keys() method. We should be overly verbose as applications
can parse the information they want independently. Prints an error
message if there's more than one key in the keyring, as well as
optionally raising a source error